Fix C4242 in ne_resource_directory NE parser
The bitwise AND 'name_info.id & 0x7FFF' promotes its uint16_t operand to
int, so the subsequent assignment to std::optional<uint16_t> triggers
C4242 (possible loss of data) on MSVC with /W3 + /WX. Cast the result
back to uint16_t explicitly.

+++ src/libexe/resources/ne_resource_directory.cpp
@@ -128,7 +128,7 @@ namespace libexe {
                     std::optional<std::string> resource_name;
 
                     if (is_integer_id) {
-                        resource_id = name_info.id & 0x7FFF;
+                        resource_id = static_cast<uint16_t>(name_info.id & 0x7FFF);
                     } else {
                         // String offset relative to beginning of resource table
                         resource_name = read_string(name_info.id);
